Cleanup Attribute Decorators
- Extracted to separate file
- Added typing to opts.key option
