+++ src/app/models/AbstractModel.ts
@@ -284,134 +284,3 @@ function createModelDecorator<S>(
     });
   };
 }
-
-/**
- * Add key to the models attributes
- */
-export function BawPersistAttr(model: AbstractModel, key: string) {
-  if (!model[AbstractModel.attributeKey]) {
-    model[AbstractModel.attributeKey] = [];
-  }
-
-  model[AbstractModel.attributeKey].push(key);
-}
-
-/**
- * Convert a collection of ids into a set
- */
-export const BawCollection = createDecorator((model, key, ids: Id[] | Ids) => {
-  if (ids instanceof Set) {
-    return;
-  }
-
-  model[key] = new Set(ids || []);
-});
-
-/**
- * Convert timestamp string into DateTimeTimezone
- */
-export const BawDateTime = createDecorator(
-  (model, key, timestamp: string | DateTime) => {
-    if (timestamp instanceof DateTime) {
-      return;
-    }
-
-    model[key] = timestamp
-      ? DateTime.fromISO(timestamp, { setZone: true })
-      : null;
-  }
-);
-
-/**
- * Convert duration string into Duration
- */
-export const BawDuration = createDecorator(
-  (model, key, seconds: number | Duration) => {
-    if (seconds instanceof Duration) {
-      return;
-    }
-
-    /*
-    Extra object fields required, do not remove. Duration calculates itself
-    based on the time spans provided, if years is removed for example,
-    the output will just keep incrementing months (i.e 24 months, instead of 2 years).
-  */
-    model[key] = seconds
-      ? Duration.fromObject({
-          years: 0,
-          months: 0,
-          days: 0,
-          hours: 0,
-          minutes: 0,
-          seconds,
-        }).normalize() // Normalize seconds into other keys (i.e 200 seconds => 3 minutes, 20 seconds)
-      : null;
-  }
-);
-
-/**
- * Abstract code required for baw decorators
- * @param setValue Set the value of the models decorated key
- */
-function createDecorator(
-  setValue: (model: AbstractModel, key: symbol, ...args: any[]) => void
-) {
-  return (opts?: BawDecoratorOptions) => (
-    model: AbstractModel,
-    key: string
-  ) => {
-    // Store decorated keys value
-    const decoratedKey = Symbol("_" + key);
-
-    Object.defineProperty(model, key, {
-      get() {
-        // If model was never set, it must be undefined
-        if (this[decoratedKey] === undefined) {
-          setValue(this, decoratedKey, undefined);
-        }
-
-        return this[decoratedKey];
-      },
-      set(args: any) {
-        // Ignore setter method if key reads value from override key
-        if (opts?.key) {
-          return;
-        }
-        setValue(this, decoratedKey, args);
-      },
-    });
-
-    // If override key provided, intercept its getter to update the decorated key
-    if (opts?.key) {
-      // Store override keys value
-      const overrideKey = Symbol("_" + opts.key);
-
-      Object.defineProperty(model, opts.key, {
-        get() {
-          return this[overrideKey];
-        },
-        set(args: any) {
-          // Update override key and set decorated key
-          this[overrideKey] = args;
-          setValue(this, decoratedKey, args);
-        },
-      });
-    }
-
-    // Should key persist in toJSON method
-    if (opts?.persist) {
-      BawPersistAttr(model, key);
-    }
-  };
-}
-
-interface BawDecoratorOptions {
-  /**
-   * Persist key in models toJSON() method
-   */
-  persist?: boolean;
-  /**
-   * Override key to read field data from another field
-   */
-  key?: string;
-}
