+++ src/app/services/baw-api/security/security.service.ts
@@ -2,7 +2,8 @@ import { HttpClient } from "@angular/common/http";
 import { Inject, Injectable, Injector } from "@angular/core";
 import { API_ROOT } from "@helpers/app-initializer/app-initializer";
 import { stringTemplate } from "@helpers/stringTemplate/stringTemplate";
-import { AbstractModel, BawPersistAttr } from "@models/AbstractModel";
+import { AbstractModel } from "@models/AbstractModel";
+import { BawPersistAttr } from "@models/AttributeDecorators";
 import { SessionUser } from "@models/User";
 import { BehaviorSubject, Observable, ObservableInput, throwError } from "rxjs";
 import { catchError, flatMap, map } from "rxjs/operators";
